changing data types

+++ Core/Entities/Concrete/UserTaxCalculation.cs
@@ -6,8 +6,8 @@
         public string EmailAddress { get; set; }
         public DateTime CalculationDate { get; set; }
         public string PostalCode { get; set; }
-        public decimal TaxCalculation { get; set; }
-        public decimal AnnualIncome { get; set; }
+        public double  TaxCalculation { get; set; }
+        public double AnnualIncome { get; set; }
